requirements.txt
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
structlog>=24.0.0
